A primitive socket server for notification system
