+++ src/lib/events.ts
@@ -1,0 +1,5 @@
+import { EventEmitter } from "events";
+
+const eventsEmit = new EventEmitter();
+
+export { eventsEmit };
